refactor: interface 위치 변경
